demo: Move egui settings/inspection windows to backend panel
